Conflicts:
	src/d_main.cpp
	src/posix/sdl/sdlglvideo.cpp
	src/rendering/hwrenderer/models/hw_models.cpp
	wadsrc/static/menudef.txt
	wadsrc/static/shaders/glsl/depthblur.fp

+++ src/rendering/2d/v_2ddrawer.cpp
@@ -643,7 +643,7 @@ void F2DDrawer::AddColorOnlyQuad(int x1, int y1, int w, int h, PalEntry color, F
 //
 //==========================================================================
 
-void F2DDrawer::AddLine(int x1, int y1, int x2, int y2, int palcolor, uint32_t color, uint8_t alpha)
+void F2DDrawer::AddLine(double x1, double y1, double x2, double y2, int palcolor, uint32_t color, uint8_t alpha)
 {
 	PalEntry p = color ? (PalEntry)color : GPalette.BaseColors[palcolor];
 	p.a = alpha;
